moved scroll function to components folder
